requirements.txt
flask>=3.0
ultralytics>=8.2.0
pillow>=10.0.0
gunicorn>=21.2
